+++ src/test/java/org/openlmis/requisition/service/BaseCommunicationServiceTest.java
@@ -27,12 +27,14 @@ import static org.junit.Assert.assertThat;
 import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.atLeastOnce;
+import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.google.common.collect.ImmutableList;
+
 import org.apache.http.NameValuePair;
 import org.junit.After;
 import org.junit.Before;
@@ -46,16 +48,18 @@ import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 import org.openlmis.requisition.testutils.DtoGenerator;
 import org.openlmis.requisition.utils.DynamicPageTypeReference;
-import org.springframework.data.domain.Page;
+import org.openlmis.requisition.utils.PageImplRepresentation;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.test.util.ReflectionTestUtils;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.HttpStatusCodeException;
 import org.springframework.web.client.RestTemplate;
+
 import java.net.URI;
 import java.util.List;
 import java.util.UUID;
@@ -226,6 +230,8 @@ public abstract class BaseCommunicationServiceTest<T> {
     service.setRestTemplate(restTemplate);
     service.setAuthService(authService);
 
+    ReflectionTestUtils.setField(service, "maxUrlLength", 2000);
+
     return service;
   }
 
@@ -262,14 +268,15 @@ public abstract class BaseCommunicationServiceTest<T> {
   }
 
   protected void mockPageResponseEntity(Object dto) {
-    ResponseEntity<Page<T>> response = stubRestTemplateAndGetPageResponseEntity();
+    ResponseEntity<PageImplRepresentation<T>> response = stubRestTemplateAndGetPageResponseEntity();
 
-    when(response.getBody())
-        .thenReturn((Page<T>) new PageImpl<>(ImmutableList.of(dto)));
+    doReturn(new PageImplRepresentation<>(new PageImpl<>(ImmutableList.of(dto))))
+        .when(response)
+        .getBody();
   }
 
-  private ResponseEntity<Page<T>> stubRestTemplateAndGetPageResponseEntity() {
-    ResponseEntity<Page<T>> response = mock(ResponseEntity.class);
+  private ResponseEntity<PageImplRepresentation<T>> stubRestTemplateAndGetPageResponseEntity() {
+    ResponseEntity<PageImplRepresentation<T>> response = mock(ResponseEntity.class);
     when(restTemplate.exchange(
         any(URI.class),
         any(HttpMethod.class),